// packages/rgs-fetcher/src/schema.ts
export interface paths {
	'/wallet/authenticate': {
		/**
		 * Authenticate Player Request
		 * @description This endpoint validates a sessionID with the operator to ensure it is a valid session.
		 */
		post: operations['authenticate'];
	};
	'/wallet/balance': {
		/**
		 * Player Balance Request
		 * @description Additional API call to receive balance of the player. Most API calls return the players balance, although a provider may want to periodically check if the balance has changed.
		 */
		post: operations['balance'];
	};
	'/wallet/play': {
		/**
		 * Play Game Request
		 * @description This API endpoint triggers a play where the Player is debited money to play a round of the game.
		 */
		post: operations['play'];
	};
	'/wallet/end-round': {
		/**
		 * Player Round End
		 * @description This API endpoint triggers an end round to occur. No futher actions can be made for this round. A payout request will be sent to the operator.
		 */
		post: operations['end-round'];
	};
	'/bet/event': {
		/**
		 * Bet Event
		 * @description This endpoint tracks where in the bet the user is up to. Send a 'event' string to this endpoint to be saved to the RGS database with the bet.
		 */
		post: operations['event'];
	};
	'/bet/action': {
		/**
		 * Bet Action
		 * @description Some games require additional actions, this endpoint will trigger an addtional action to occur if the game has been setup for it.
		 */
		post: operations['action'];
	};
	'/session/start': {
		/**
		 * Start Session
		 * @description This endpoint is must be created on the RGS side that an operator can send a request to. The operator will send a request to this endpoint to get a URL for the game that the player wishes to play. **Please send this API request to the mock-casino endpoint of the environment.**
		 */
		post: operations['session_start'];
	};
	'/game/search': {
		/**
		 * @description TBD
		 */
		post: operations['search'];
	};
}

export type webhooks = Record<string, never>;

export interface components {
	schemas: {
		Error: any;
		Search: {
			bookID?: number;
			kind?: number;
			symbol?: string;
			hasWild?: boolean;
			wildMult?: number;
			gameType?: string;
		};
		req_search: {
			mode?: components['schemas']['Mode'];
			search?: components['schemas']['Search'];
		};
		res_search: {
			balance?: components['schemas']['BalanceObject'];
			round?: components['schemas']['RoundDetailObject'];
			error?: components['schemas']['Error'];
		};
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}

export type external = Record<string, never>;

export interface operations {
	/**
	 * @description TBD
	 */
	search: {
		/** @description TBD **/
		requestBody: {
			content: {
				'application/json': components['schemas']['req_search'];
			};
		};
		responses: {
			/** @description OK */
			200: {
				content: {
					'application/json': components['schemas']['res_search'];
				};
			};
		};
	};
	/**
	 * Authenticate Player Request
	 * @description This endpoint validates a sessionID with the operator to ensure it is a valid session.
	 */
	authenticate: {
		/** @description Session Information */
		requestBody: {
			content: {
				'application/json': components['schemas']['req_authenticate'];
			};
		};
		responses: {
			/** @description OK */
			200: {
				content: {
					'application/json': components['schemas']['res_authenticate'];
				};
			};
		};
	};
	/**
	 * Player Balance Request
	 * @description Additional API call to receive balance of the player. Most API calls return the players balance, although a provider may want to periodically check if the balance has changed.
	 */
	balance: {
		/** @description Session Information */
		requestBody: {
			content: {
				'application/json': components['schemas']['req_Balance'];
			};
		};
		responses: {
			/** @description OK */
			200: {
				content: {
					'application/json': components['schemas']['res_Balance'];
				};
			};
		};
	};
	/**
	 * Play Game Request
	 * @description This API endpoint triggers a play where the Player is debited money to play a round of the game.
	 */
	play: {
		/** @description Session Information */
		requestBody: {
			content: {
				'application/json': components['schemas']['req_play'];
			};
		};
		responses: {
			/** @description OK */
			200: {
				content: {
					'application/json': components['schemas']['res_play'];
				};
			};
		};
	};
	/**
	 * Player Round End
	 * @description This API endpoint triggers an end round to occur. No futher actions can be made for this round. A payout request will be sent to the operator.
	 */
	'end-round': {
		/** @description Session Information */
		requestBody: {
			content: {
				'application/json': components['schemas']['req_end_round'];
			};
		};
		responses: {
			/** @description OK */
			200: {
				content: {
					'application/json': components['schemas']['res_end_round'];
				};
			};
		};
	};
	/**
	 * Bet Event
	 * @description This endpoint tracks where in the bet the user is up to. Send a 'event' string to this endpoint to be saved to the RGS database with the bet.
	 */
	event: {
		/** @description Session Information */
		requestBody: {
			content: {
				'application/json': components['schemas']['req_event'];
			};
		};
		responses: {
			/** @description OK */
			200: {
				content: {
					'application/json': components['schemas']['res_event'];
				};
			};
		};
	};
	/**
	 * Bet Action
	 * @description Some games require additional actions, this endpoint will trigger an addtional action to occur if the game has been setup for it.
	 */
	action: {
		/** @description Session Information */
		requestBody: {
			content: {
				'application/json': components['schemas']['req_action'];
			};
		};
		responses: {
			/** @description OK */
			200: {
				content: {
					'application/json': components['schemas']['res_action'];
				};
			};
		};
	};
	/**
	 * Start Session
	 * @description This endpoint is must be created on the RGS side that an operator can send a request to. The operator will send a request to this endpoint to get a URL for the game that the player wishes to play. **Please send this API request to the mock-casino endpoint of the environment.**
	 */
	session_start: {
		/** @description Session Information */
		requestBody: {
			content: {
				'application/json': components['schemas']['req_sess_start'];
			};
		};
		responses: {
			/** @description OK */
			200: {
				content: {
					'application/json': components['schemas']['res_sess_start'];
				};
			};
		};
	};
}